requirements.txt
wikipedia-api~=0.5.4
pyttsx3~=2.90
vosk~=0.3.7
termcolor~=1.1.0
sounddevice
soundfile
pyautogui
requests
numpy
audioplayer
python-dateutil~=2.6.0
pyyaml~=6.0
